[hoyeong.jin] feat: 설계, 구조 설명 보충,
